refactor(app): share preview diff generation
