+++ openspec/changes/refactor-app-preview-diff/README.md
@@ -1,0 +1,3 @@
+# refactor-app-preview-diff
+
+Refactor: extract shared preview diff logic into app layer
