+++ src/app/mod.rs
@@ -1,0 +1,1 @@
+pub(crate) mod preview_diff;
